Change gain-pan-stereo to mix-utility. Added polarity to both left and right channels. Fixed the display values for pan parameter.
