+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
 members = [
-    "gain-pan-stereo"
+    "mix-utility"
 ]
