--- gain-pan-stereo/README.md
@@ -1,1 +1,0 @@
-cargo xtask bundle gain-pan-stereo --release
