+++ mix-utility/README.md
@@ -1,0 +1,1 @@
+cargo xtask bundle mix-utility --release
